add base mock servers for each command, fix creation of mock server, edit test files for stream and the files itself
